Rename plural package to poplural
