+++ internal/poplural/parser_test.go
@@ -1,4 +1,4 @@
-package plural
+package poplural
 
 import (
 	"encoding/json"
