+++ internal/poplural/table_test.go
@@ -1,4 +1,4 @@
-package plural
+package poplural
 
 import (
 	"reflect"
